chunkers/requirements.txt
grpcio
grpcio-tools
grpcio-health-checking
grpcio-reflection
PyYAML
langchain-text-splitters
tiktoken